Hide Access Controls links/pages based on scope
See #11383

Ensure that no visitor to the Teleport docs site sees content that is
irrelevant to their scope (e.g., Cloud, Open Source, or Enterprise) by
hiding scope-irrelevant content from the navigation menu and menu
pages.

For pages that are only relevant to a specific scope, show users with
unintended scopes a menu of links to supported scopes.

This PR focuses on the Access Controls section.